+++ Decision System/Core/Models/GroupOfGroups.cs
@@ -8,10 +8,10 @@ namespace Core.Models
 {
     public class GroupOfGroups : Group
     {
-        private List<int> GroupIds;
+        private List<long> GroupIds;
 
         public virtual List<Group> Groups { get; set; }
-        public List<int> GroupIds1 { get => GroupIds; set => GroupIds = value; }
+        public List<long> GroupIds1 { get => GroupIds; set => GroupIds = value; }
 
         public GroupOfGroups()
         {
